requirements.txt
langchain>=0.3.0
langchain-nvidia-ai-endpoints>=0.3.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
ddgs>=0.1.0
